[M, F] minor changes

+++ Sources/FusionKit/Model/FusionMessage.swift
@@ -19,7 +19,7 @@ public protocol FusionMessage: Sendable { }
 /// The `FusionFrame` protocol for message conformance
 protocol FusionFrame: FusionMessage {
     var opcode: UInt8 { get }
-    var size: UInt32 { get }
+    var size: UInt64 { get }
     var encode: Data { get }
     static func decode(from payload: Data) -> FusionFrame?
 }
@@ -29,7 +29,7 @@ protocol FusionFrame: FusionMessage {
 /// Conformance to protocol `FusionFrame` and `FusionMessage`
 extension UInt16: FusionFrame {
     var opcode: UInt8 { FusionOpcode.uint16.rawValue }
-    var size: UInt32 { UInt32(self.encode.count + FusionPacket.header.rawValue) }
+    var size: UInt64 { UInt64(self.encode.count + FusionPacket.header.rawValue) }
     var encode: Data { Data(count: Int(self)) }
     static func decode(from payload: Data) -> FusionFrame? { Self(payload.count) }
 }
@@ -37,7 +37,7 @@ extension UInt16: FusionFrame {
 /// Conformance to protocol `FusionFrame` and `FusionMessage`
 extension String: FusionFrame {
     var opcode: UInt8 { FusionOpcode.string.rawValue }
-    var size: UInt32 { UInt32(self.encode.count + FusionPacket.header.rawValue) }
+    var size: UInt64 { UInt64(self.encode.count + FusionPacket.header.rawValue) }
     var encode: Data { Data(self.utf8) }
     static func decode(from payload: Data) -> FusionFrame? { Self(bytes: payload, encoding: .utf8) }
 }
@@ -45,7 +45,7 @@ extension String: FusionFrame {
 /// Conformance to protocol `FusionFrame` and `FusionMessage`
 extension Data: FusionFrame {
     var opcode: UInt8 { FusionOpcode.data.rawValue }
-    var size: UInt32 { UInt32(self.encode.count + FusionPacket.header.rawValue) }
+    var size: UInt64 { UInt64(self.encode.count + FusionPacket.header.rawValue) }
     var encode: Data { self }
     static func decode(from payload: Data) -> FusionFrame? { payload }
 }
